requirements.txt
dj-database-url==0.4.1
Django==1.9.5
django-bootstrap==0.2.4
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-rest-auth==0.7.0
django-rest-swagger==0.3.5
djangorestframework==3.3.3
gunicorn==19.4.5
Pillow==3.2.0
psycopg2==2.6.1
PyYAML==3.11
six==1.10.0
sqlparse==0.1.19
whitenoise==3.0
